# Convince the compiler to auto-vectorize the range check in parquet DictionaryBuffer